regard match message approach as one pipe to fetch fingerprints
